requirements.txt
django
djangorestframework
environs
flake8
flake8-django
django-cors-headers
